==> 00.sample-dualindexing.ipynb <==
## Dual Indexing-Sample Layout 

import pandas as pd
import math
from datetime import  datetime

dt = datetime.today().strftime(format='%d-%m-%Y')

### Load seq cheat sheet sample layout

***Samples must be in multiple of 96 per plate, if not extrapolate arbitrarily to fill (a) plate(s)***

***The columns for sample `name/id` and `plate-position` must be provided***

parent_dir = '/home/kibet/bioinformatics/github/covid/data/SarsGenomics/SARS-CoV-2/SeqSampleSheets'
seq = 'seq53'
df_seqcheat = (pd.read_excel(f'{parent_dir}/seq53_sampleSheet_JJ.xlsx', 
                usecols=['s_num', 'idx_plt_pos'], header=0).
               dropna())

samp_list = list(df_seqcheat['s_num'])

len(samp_list)

count = 0
plates = []
plate = {'col1': [], 'col2': [], 'col3': [], 'col4': [], 'col5': [], 'col6': [],
         'col7': [], 'col8': [], 'col9': [], 'col10': [], 'col11': [], 'col12': []}
if len(samp_list) == 384:
    nrows = 32
elif len(samp_list) == 96:
    nrows = 8
else: 
    print('The number of samples is not 96 or 384\nSamples must be in multiple of 96 per plate, if not extrapolate arbitrarily to fill (a) plate(s)')
for r in range(nrows):
    row = samp_list[count:count+12]
    plate['col1'].append(row[0])
    plate['col2'].append(row[1])
    plate['col3'].append(row[2])
    plate['col4'].append(row[3])
    plate['col5'].append(row[4])
    plate['col6'].append(row[5])
    plate['col7'].append(row[6])
    plate['col8'].append(row[7])
    plate['col9'].append(row[8])
    plate['col10'].append(row[9])
    plate['col11'].append(row[10])
    plate['col12'].append(row[11])
    count += 12

a = 8
plt1=[]; plt2=[]; plt3=[]; plt4=[]
for n in list(plate.values()):
    if nrows == 8:
        plt1 += n[:a]
        all_plates = plt1
    elif nrows == 32:
        plt1 += n[:a]
        plt2 += n[a:a*2]
        plt3 += n[a*2:a*3]
        plt4 += n[a*3:]
        all_plates = plt1 + plt2 + plt3 + plt4
    else: print('The number of plates is not 1 or 4')

sr_index = pd.Series(all_plates, name='indexing')
len(sr_index)

df_index = sr_index.to_frame().merge(df_seqcheat, how='inner', left_on='indexing', right_on='s_num')

df_index

df_index.to_excel(f'{parent_dir}/index_cheat_sheet_{seq}.xlsx', index=False)
