==> 02.metadata-cleaning.ipynb <==
$29^{th}$ Aug 2021

import pandas as pd
import seaborn as sns
import numpy as np
import glob, os, re
import warnings
import matplotlib.pyplot as plt
from datetime import  datetime
from ipywidgets import widgets, interactive

dt = datetime.today().strftime(format='%d-%m-%Y')

### **Functions**

def replace(string, substitutions):
    substrings = sorted(substitutions, key=len, reverse=True)
    regex = re.compile('|'.join(map(re.escape, substrings)))
    return regex.sub(lambda match: substitutions[match.group(0)], string)

def rename_cols_subset_df(df, cols):
    df_new = df.copy()
    df_new.columns = list(map(lambda x: x.split('(')[0].strip(' _'), [col.replace(' ', '_').replace('\n', '_').upper() for col in df.columns]))
    df_cln = df_new[df_new['SAMPLE_NUMBER'].notna()]
    return df_cln[cols]

def cov_names(x):
    x = x.upper().replace(' ', '')
    if len(x) == 8:
        x = x.replace('COVC', 'COVC0')
    elif len(x) == 7:
        x = x.replace('COVC', 'COVC00')
    elif len(x) == 6:
        x = x.replace('COVC', 'COVC000')
    elif len(x) == 5:
        x = x.replace('COVC', 'COVC0000')
    return x

def my_date_parser(x):
    try:
        if type(x) == pd.Timestamp:
            return x#.strftime('%d-%m-%Y')
        if type(x) != pd.Timestamp and type(x) != (int, float, pd.NaT):
            return pd.to_datetime(x, errors='coerce', dayfirst=True)#.strftime('%d-%m-%Y')
        if x in ['None indicated', 'NIL', 'Not indicated on form ',
                    'Nil', 'None indicated', 'Leaked Sample - Empty', 'nan']:
            return pd.NaT
    except (ValueError, AttributeError):
        return pd.NaT

def cln_results(x):
    if isinstance(x, str):
        x = x.strip().lower()
        if 'neg' in x:
            x = x.replace(x, 'Negative')
        elif 'pos' in x:
            x = x.replace(x, 'Positive')
        elif x == 'inconclusive':
            x = x.replace(x, 'Negative')
    return x
        

### **Establish parent directory**

# uniqueness in directory and file names is assumed for all analyses
sars_dir = "bioinformatics/github/covid/data/SarsGenomics/SARS-CoV-2"
home_dir = os.getenv('HOME')
parent_dir = glob.glob(f'{home_dir}/{sars_dir}', recursive=True)[0]

cols_of_interest = ['CASE_ID','SAMPLE_NUMBER', 'NAME', 'AGE', 'AGE_UNIT', 'GENDER', 'OCCUPATION', 'NATIONALITY'
           , 'COUNTY_OF_RESIDENCE', 'SUB_COUNTY_OF_RESIDENCE', 'HAS_TRAVEL_HOSTORY', 'TRAVEL_FROM'
           , 'QUARANTINE_FACILITY/HOSPITAL/HOMESTEAD', 'SYMTOMS_SHOWN', 'SAMPLE_TYPE', 'DATE_OF_SAMPLE_COLLECTION'
           , 'DATE_SAMPLE_RECEIVED_IN_THE_LAB', 'RESULT', 'LAB_CONFIRMATION_DATE', 'VACCINATION_STATUS_Y/N'
           , 'DOSAGE_C/NC']

files = os.listdir(glob.glob(f'{parent_dir}/**/ResultsDBCurated', recursive=True)[0])
files = sorted(files)

# keyerrors = []
# df_raw = pd.DataFrame()
# count = 0
# for file in files:
#     try:
#         df_file = pd.read_excel(f'{data_dir}/{file}', header=1)
#         df_data = rename_cols_subset_df(df_file, 'DATE_SAMPLE_RECEIVED_IN_THE_LAB')
# #         df_raw = df_raw.append(df_data)
#         count += 1
#     except: 
#         if KeyError:
#             keyerrors.append(file)
#         elif ValueError:
#             print(f'ValueError: {file}')
#     finally: pass
# print(len(keyerrors))
# sorted(keyerrors)

# files = ['COVID-19-Test_result_Reporting_01_07_2020-A.xlsx',
#  'COVID-19-Test_result_Reporting_01_07_2020-B.xlsx',
#  'COVID-19-Test_result_Reporting_05_09_2020-A.xlsx',
#  'COVID-19-Test_result_Reporting_05_09_2020-B.xlsx',
#  'COVID-19-Test_result_Reporting_08_09_2020.xlsx',
#  'COVID-19-Test_result_Reporting_15_07_2020.xlsx',
#  'COVID-19-Test_result_Reporting_18_06_2020.xlsx',
#  'COVID-19-Test_result_Reporting_31_07_2020-A.xlsx',
#  'COVID-19-Test_result_Reporting_31_07_2020-B.xlsx']

# for file in files:
#     df = pd.read_excel(f'{data_dir}/{file}', header=1)
#     try:
#         df.insert(27, 'DATE SAMPLE RECEIVED IN THE LAB', pd.NaT)
#         df.to_excel(f'/home/douso/Documents/TrendData/Results/add_col/{file}', index=False, startrow=1)
#     except:
#         raise
#         print (f'Error: {file}')
#     finally: pass

# keyerrors = []
# df_raw = pd.DataFrame()
# count = 0
# for file in files:
#     try:
#         df_file = pd.read_excel(f'{data_dir}/{file}', header=1)
#         df_data = rename_cols_subset_df(df_file, cols_of_interest)
#         df_raw = df_raw.append(df_data)
#         count += 1
#     except: 
#         if KeyError:
#             keyerrors.append(file)
#         elif ValueError:
#             print(f'ValueError: {file}')
#     finally: pass
# print(len(keyerrors))
# sorted(keyerrors)

headers = ['TESTING LAB', 'CASE ID', 'TYPE OF CASE'
       , 'SAMPLE NUMBER', 'NAME', 'ID/PASSPORT NUMBER', 'AGE'
       , 'AGE UNIT', 'GENDER', 'PHONE NUMBER'
       , 'OCCUPATION', 'NATIONALITY', 'COUNTY OF RESIDENCE'
       , 'SUB COUNTY OF RESIDENCE', 'VILLAGE/ESTATE OF RESIDENCE', 'WARD'
       , 'COUNTY OF DIAGNOSIS', 'HAS TRAVEL HOSTORY'
       , 'TRAVEL FROM', 'CONTACT WITH CASE Y/N', 'CONFIRMED CASE NAME'
       , 'QUARANTINE FACILITY/HOSPITAL/HOMESTEAD', 'HAVE SYMPTOMS Y/N'
       , 'DATE OF ONSET OF SYMPTOMS', 'SYMTOMS SHOWN'
       , 'SAMPLE TYPE', 'DATE OF SAMPLE COLLECTION', 'DATE SAMPLE RECEIVED IN THE LAB'
       , 'RESULT', 'LAB CONFIRMATION DATE', 'VACCINATION STATUS Y/N', 'DOSAGE C/NC']

id_headers = ['SAMPLE_NUMBER', 'NAME', 'ID/PASSPORT_NUMBER', 'AGE',
       'AGE_UNIT', 'GENDER', 'PHONE_NUMBER', 'RESULT',
       'LAB_CONFIRMATION_DATE']

# for file in keyerrors[:-1]:
#     df_file = pd.read_excel(f'{data_dir}/{file}', header=1)
#     df_file.columns = headers
#     df_file.to_excel(f'{data_dir}/add_col2/{file}', startrow=1)

The following 60 sheets, besides some June sheets (manually fixed earlier), lacked the `DATE_SAMPLE_RECEIVED_IN_THE_LAB` column, it was added for uniformity:

- 'COVID-19-Test_result_Reporting_01_07_2020.xlsx'
- 'COVID-19-Test_result_Reporting_01_09_2020.xlsx'
- 'COVID-19-Test_result_Reporting_02_09_2020.xlsx'
- 'COVID-19-Test_result_Reporting_03_07_2020.xlsx'
- 'COVID-19-Test_result_Reporting_03_08_2020.xlsx'
- 'COVID-19-Test_result_Reporting_03_09_2020.xlsx'
- 'COVID-19-Test_result_Reporting_04_07_2020.xlsx'
- 'COVID-19-Test_result_Reporting_04_08_2020.xlsx'
- 'COVID-19-Test_result_Reporting_04_09_2020.xlsx'
- 'COVID-19-Test_result_Reporting_05_07_2020.xlsx'
- 'COVID-19-Test_result_Reporting_05_08_2020.xlsx'
- 'COVID-19-Test_result_Reporting_05_09_2020.xlsx'
- 'COVID-19-Test_result_Reporting_06_07_2020.xlsx'
- 'COVID-19-Test_result_Reporting_06_08_2020.xlsx'
- 'COVID-19-Test_result_Reporting_07_07_2020.xlsx'
- 'COVID-19-Test_result_Reporting_07_08_2020.xlsx'
- 'COVID-19-Test_result_Reporting_07_09_2020.xlsx'
- 'COVID-19-Test_result_Reporting_08_07_2020.xlsx'
- 'COVID-19-Test_result_Reporting_08_08_2020.xlsx'
- 'COVID-19-Test_result_Reporting_09_07_2020.xlsx'
- 'COVID-19-Test_result_Reporting_09_08_2020.xlsx'
- 'COVID-19-Test_result_Reporting_10_08_2020.xlsx'
- 'COVID-19-Test_result_Reporting_11_08_2020.xlsx'
- 'COVID-19-Test_result_Reporting_12_08_2020.xlsx'
- 'COVID-19-Test_result_Reporting_13_07_2020.xlsx'
- 'COVID-19-Test_result_Reporting_13_08_2020.xlsx'
- 'COVID-19-Test_result_Reporting_14_07_2020.xlsx'
- 'COVID-19-Test_result_Reporting_14_08_2020.xlsx'
- 'COVID-19-Test_result_Reporting_15_08_2020-A.xlsx'
- 'COVID-19-Test_result_Reporting_15_08_2020-B.xlsx'
- 'COVID-19-Test_result_Reporting_16_06_2020.xlsx'
- 'COVID-19-Test_result_Reporting_16_10_2020 (1).xlsx'
- 'COVID-19-Test_result_Reporting_17_07_2020.xlsx'
- 'COVID-19-Test_result_Reporting_17_08_2020.xlsx'
- 'COVID-19-Test_result_Reporting_18_07_2020.xlsx'
- 'COVID-19-Test_result_Reporting_18_08_2020.xlsx'
- 'COVID-19-Test_result_Reporting_19_06_2020.xlsx'
- 'COVID-19-Test_result_Reporting_19_08_2020.xlsx'
- 'COVID-19-Test_result_Reporting_20_06_2020.xlsx'
- 'COVID-19-Test_result_Reporting_20_08_2020.xlsx'
- 'COVID-19-Test_result_Reporting_21_06_2020.xlsx'
- 'COVID-19-Test_result_Reporting_21_07_2020.xlsx'
- 'COVID-19-Test_result_Reporting_21_08_2020.xlsx'
- 'COVID-19-Test_result_Reporting_22_08_2020.xlsx'
- 'COVID-19-Test_result_Reporting_23_07_2020.xlsx'
- 'COVID-19-Test_result_Reporting_24_06_2020.xlsx'
- 'COVID-19-Test_result_Reporting_24_07_2020.xlsx'
- 'COVID-19-Test_result_Reporting_24_08_2020-A.xlsx'
- 'COVID-19-Test_result_Reporting_24_08_2020-B.xlsx'
- 'COVID-19-Test_result_Reporting_25_06_2020.xlsx'
- 'COVID-19-Test_result_Reporting_25_08_2020.xlsx'
- 'COVID-19-Test_result_Reporting_27_06_2020.xlsx'
- 'COVID-19-Test_result_Reporting_27_07_2020.xlsx'
- 'COVID-19-Test_result_Reporting_27_08_2020.xlsx'
- 'COVID-19-Test_result_Reporting_28_06_2020.xlsx'
- 'COVID-19-Test_result_Reporting_28_07_2020.xlsx'
- 'COVID-19-Test_result_Reporting_28_08_2020.xlsx'
- 'COVID-19-Test_result_Reporting_29_06_2020.xlsx'
- 'COVID-19-Test_result_Reporting_29_07_2020.xlsx'
- 'COVID-19-Test_result_Reporting_29_08_2020.xlsx'

# for file in files:
#     df = pd.read_excel(f'{data_dir}/{file}', header=1)
#     try:
#         df.insert(27, 'DATE SAMPLE RECEIVED IN THE LAB', np.nan)
#         df.to_excel(f'{data_dir}/add_col/{file}', index=False, startrow=1)
#     except:
#         raise
#         print (f'Error: {file}')
#     finally: pass

The follwing sheets have missing values for the `RESULT` column:
- 'COVID-19-Test_result_Reporting_20_06_2020.xlsx'
- 'COVID-19-Test_result_Reporting_16_06_2020.xlsx'

Updated column names of some of the Jun files ['GENDER_(M/F)', 'LAB_CONFIRMATION_DATE', 'RESULT']

# df_raw = pd.DataFrame()
# count = 0
# for file in files:
#     if file.endswith('.xlsx'):
#         try:
#             df_file = pd.read_excel(f'{data_dir}/{file}', header=1)
#             df_data = rename_cols_subset_df(df_file, cols_of_interest)
#             df_raw = df_raw.append(df_data)
#             count += 1
#         except: 
#             if KeyError:
# #                 raise
#                 print(f'KeyError: {file}')
#             elif ValueError:
#                 print(f'ValueError: {file}')
#         finally: pass
#     pass
# print(f'Files processed: {count}')
# print(f'Files in the folder: {len(files)}')

### Travel History Missing - the HOSTORY misspelling

- COVID-19-Test_result_Reporting_09_09_2020.xlsx
- COVID-19-Test_result_Reporting_11_09_2020.xlsx
- COVID-19-Test_result_Reporting_12_09_2020.xlsx
- COVID-19-Test_result_Reporting_14_09_2020.xlsx
- COVID-19-Test_result_Reporting_15_09_2020.xlsx
- COVID-19-Test_result_Reporting_18_10_2020.xlsx
- Files processed: 249
- Files in the folder: 256

# keyerrors = []
# # df_raw = pd.DataFrame()
# count = 0
# for file in files:
#     try:
#         df_file = pd.read_excel(f'{data_dir}/{file}', header=1)
#         df_data = rename_cols_subset_df(df_file, cols_of_interest)
# #         df_raw = df_raw.append(df_data)
#         count += 1
#     except: 
#         if KeyError:
#             keyerrors.append(file)
#         elif ValueError:
#             print(f'ValueError: {file}')
#     finally: pass
# print(len(keyerrors))
# sorted(keyerrors)

# files = ['COVID-19-Test_result_Reporting_09_09_2020.xlsx',
#  'COVID-19-Test_result_Reporting_11_09_2020.xlsx',
#  'COVID-19-Test_result_Reporting_12_09_2020.xlsx',
#  'COVID-19-Test_result_Reporting_14_09_2020.xlsx',
#  'COVID-19-Test_result_Reporting_15_09-2020.xlsx',
#  'COVID-19-Test_result_Reporting_18_10_2020.xlsx']

# for file in files:
#     df_file = pd.read_excel(f'{data_dir}/{file}', header=1)
#     (df_file.rename(columns={'HAS TRAVEL HISTORY(LAST 14 DAYS) Y/N': 'HAS TRAVEL HOSTORY(LAST 14 DAYS) Y/N'}, errors='raise').
#     to_excel(f'{data_dir}/rename_col/{file}', index=False, startrow=1))

### Start of find repeat tests

sorted(list(set(map(lambda x: x.replace('_', ' '), id_headers)) & set(headers))) == sorted(list(map(lambda x: x.replace('_', ' '), id_headers)))

df_raw_ids = pd.DataFrame()
count = 0
for file in files:
    if file.endswith('.xlsx'):
        try:
            df_file = pd.read_excel(f'{data_dir}/{file}', header=1)
            if (sorted(list(set(map(lambda x: x.replace('_', ' '), 
                id_headers)) & set(headers))) == sorted(list(map(lambda x: 
                x.replace('_', ' '), id_headers)))):
                df_data = rename_cols_subset_df(df_file, id_headers)
                df_raw_ids = df_raw_ids.append(df_data)
                count += 1
            
        except: 
            if KeyError:
                print(f'KeyError: {file}')
                raise
                print(f'KeyError: {file}')
            elif ValueError:
                print(f'ValueError: {file}')
        finally: pass
    pass
print(f'Files processed: {count}')
print(f'Files in the folder: {len(files)}')

df_raw_ids['LAB_CONFIRMATION_DATE'] = df_raw_ids['LAB_CONFIRMATION_DATE'].apply(lambda x: pd.to_datetime(x, errors='coerce'))

df_raw_ids.info()#pd.to_datetime('2021-01-01 00:00:00', errors='coerce')

df_2021 = df_raw_ids[df_raw_ids['LAB_CONFIRMATION_DATE'] > pd.to_datetime('2021-01-01')]# & df_raw_ids['LAB_CONFIRMATION_DATE'] < pd.to_datetime(datetime.now())]
df_2021.shape

df_2021['ID/PASSPORT_NUMBER']=df_2021['ID/PASSPORT_NUMBER'].astype(str).apply(lambda x: x.replace('.0', ''))

df_2021[df_2021.sort_values('ID/PASSPORT_NUMBER', ignore_index=True).duplicated(['ID/PASSPORT_NUMBER'], False)==True].sort_values('ID/PASSPORT_NUMBER').drop_duplicates(keep=False).sort_values('ID/PASSPORT_NUMBER').to_excel('../SomeFIle.xlsx')#dropna(subset=['ID/PASSPORT_NUMBER'])

pd.to_datetime(datetime.now())

df_raw_ids['PHONE_NUMBER']=df_raw_ids['PHONE_NUMBER'].apply(lambda x: str(x).lstrip('0').replace('.0', ''))

df_ids_phone = df_raw_ids[(df_raw_ids
             .sort_values('PHONE_NUMBER', ignore_index=True)#, key=lambda col: col.astype(str))
            .duplicated(['PHONE_NUMBER'], keep=False)) == True]
df_ids_phone.shape

phones = df_ids_phone.dropna(subset=['PHONE_NUMBER'])
phones.shape

phones = df_ids_phone[df_ids_phone['PHONE_NUMBER'].isna() == False].sort_values('PHONE_NUMBER', key=lambda col: col.astype(str))
phones.shape

phones.head()#['PHONE_NUMBER']=phones['PHONE_NUMBER'].astype(str).apply(lambda x: x.rstrip('.0'))

phones2 = phones.sort_values('PHONE_NUMBER').drop_duplicates('PHONE_NUMBER', False)
phones2.shape

phones2_list = list(phones2['PHONE_NUMBER'])

df_phones_rep = phones[phones['PHONE_NUMBER'].isin(phones2_list) == False].sort_values('SAMPLE_NUMBER').drop_duplicates(['SAMPLE_NUMBER'], False).sort_values('PHONE_NUMBER')
df_phones_rep = df_phones_rep[df_phones_rep.duplicated('PHONE_NUMBER', False) == True]
df_phones_rep.shape

df_phones_rep1 = df_phones_rep.drop(df_phones_rep[df_phones_rep['PHONE_NUMBER'].str.contains('ndicat')].index)

df_phones_rep2 = df_phones_rep1.drop(df_phones_rep1[df_phones_rep1['PHONE_NUMBER'].str.contains('il')].index)
df_phones_rep2.head()

df_phones_rep3 = df_phones_rep1.drop(df_phones_rep1[df_phones_rep1['PHONE_NUMBER'].str.contains('nan')].index)
df_phones_rep3

df_ids_id = df_raw_ids[(df_raw_ids
             .sort_values(['ID/PASSPORT_NUMBER'], ignore_index=True, key=lambda col: col.astype(str))
            .duplicated(['ID/PASSPORT_NUMBER'], keep=False)) == True]

ids = df_ids_id[df_ids_id['ID/PASSPORT_NUMBER'].isna() == False].sort_values('ID/PASSPORT_NUMBER', key=lambda col: col.astype(str))

ids['ID/PASSPORT_NUMBER']=ids['ID/PASSPORT_NUMBER'].astype(str).apply(lambda x: x.rstrip('.0'))

ids.head()

COVM01626	NPHL04/JKIA35539/2021	29	M	NA	South Africa	Nairobi	Nan	No	NA	NA	NA	NP & OP Swab	NA	2021-11-29 00:00:00	NA	2021-11-29 00:00:00	NA	NA	NA	NA	NA
ids2 = ids.sort_values('ID/PASSPORT_NUMBER').drop_duplicates('ID/PASSPORT_NUMBER', False)

ids2_list = list(ids2['ID/PASSPORT_NUMBER'])

df_ids_rep = ids[ids['ID/PASSPORT_NUMBER'].isin(ids2_list) == False].drop_duplicates(['SAMPLE_NUMBER'], False).sort_values('ID/PASSPORT_NUMBER')
df_ids_rep = df_ids_rep[df_ids_rep.duplicated('ID/PASSPORT_NUMBER', False) == True]

df_ids_rep1 = df_ids_rep.drop(df_ids_rep[df_ids_rep['ID/PASSPORT_NUMBER'].str.contains('ndicat')].index)

df_ids_rep2 = df_ids_rep1.drop(df_ids_rep1[df_ids_rep1['ID/PASSPORT_NUMBER'].str.contains('il')].index)
df_ids_rep2.sort_values(['ID/PASSPORT_NUMBER'])
print(df_ids_rep2.shape)
df_ids_rep2.head()

### End of find repeat tests

id_headers = ['SAMPLE_NUMBER', 'NAME', 'AGE',
       'AGE_UNIT', 'GENDER', 'RESULT',
       'LAB_CONFIRMATION_DATE']

warnings.filterwarnings('ignore', module='openpyxl')

df_raw = pd.DataFrame()
count = 0
for file in files:
    if file.endswith('.xlsx'):
        try:
            df_file = pd.read_excel(f"{glob.glob(f'{parent_dir}/**/ResultsDBCurated', recursive=True)[0]}/{file}", header=1)
            if 'VACCINATION STATUS Y/N' not in df_file.columns:
                df_file[['VACCINATION_STATUS_Y/N', 'DOSAGE_C/NC']] = [pd.NA, pd.NA]
            df_data = rename_cols_subset_df(df_file, cols_of_interest)
            df_raw = pd.concat([df_raw,df_data])
            count += 1
        except:
            if KeyError:
                print(f'KeyError: {file}')
#                raise
                print(f'KeyError: {file}')
            if ValueError:
                print(f'ValueError: {file}')
#                 raise
        finally: pass
    pass
print(f'Files processed: {count}')
print(f'Files in the folder: {len(files)}')

KeyError: COVID19_Reporting_06-10-2020.xlsx
KeyError: COVID19_Reporting_16-09-2020.xlsx
KeyError: COVID19_Reporting_17-09-2020.xlsx
KeyError: COVID19_Reporting_18-09-2020.xlsx
KeyError: COVID19_Reporting_19-09-2020.xlsx
KeyError: COVID19_Reporting_25-10-2020-A.xlsx
Files processed: 273
Files in the folder: 281

KeyError: COVID-19-Test_result_Reporting_06_10_2020.xlsx
KeyError: COVID-19-Test_result_Reporting_16_09_2020.xlsx
KeyError: COVID-19-Test_result_Reporting_17_09_2020.xlsx
KeyError: COVID-19-Test_result_Reporting_18_09_2020.xlsx
KeyError: COVID-19-Test_result_Reporting_19_09_2020.xlsx
KeyError: COVID-19-Test_result_Reporting_25_10_2020-B.xlsx
KeyError: COVID19_Reporting_06-10-2020.xlsx
KeyError: COVID19_Reporting_16-09-2020.xlsx
KeyError: COVID19_Reporting_17-09-2020.xlsx
KeyError: COVID19_Reporting_18-09-2020.xlsx
KeyError: COVID19_Reporting_19-09-2020.xlsx
KeyError: COVID19_Reporting_24-10-2020-A.xlsx
KeyError: COVID19_Reporting_25-10-2020-A.xlsx
KeyError: COVID19_Reporting_27-10-2020.xlsx
Files processed: 265
Files in the folder: 281

# df_raw.head()

### **Clean Headers**

cols_rename = ['CASE_ID', 'S_NUM', 'AGE', 'AGE_UNIT', 'GEND', 'OCCU', 'NAT', 'COUNT_RES', 'S_COUNT_RES', 'TRAV_HIST', 
                    'TRAV_FROM', 'QUAR_PLACE', 'SYMPS', 'SAMP_TYPE', 'DT_SAM_COLL', 'DT_SAM_RECEP', 'RESULT', 'DT_CONF', 'VACC_STATUS', 'VACC_DOS']

# df_raw.columns = cols_rename

### **Clean Sample Numbers**

cols_of_interest2 = ['CASE_ID', 'S_NUM', 'NAME', 'AGE', 'AGE_UNIT', 'GEND', 'OCCU', 'NAT', 'COUNT_RES', 'S_COUNT_RES', 'TRAV_HIST', 
                    'TRAV_FROM', 'QUAR_PLACE', 'SYMPS', 'SAMP_TYPE', 'DT_SAM_COLL', 'DT_SAM_RECEP', 'RESULT', 'DT_CONF', 'VACC_STATUS', 'VACC_DOS']

df_raw.to_excel(f"{glob.glob(f'{parent_dir}/**/Outputs', recursive=True)[0]}/COVID19-results-merged.xlsx", 
              index=False, header=cols_of_interest2)

df_mrg_mtdata = pd.read_excel(f"{glob.glob(f'{parent_dir}/**/Outputs', recursive=True)[0]}/COVID19-results-merged.xlsx", usecols=cols_rename)

### **Clean Sample Name**

df_name = df_mrg_mtdata.assign(S_NUM=df_mrg_mtdata['S_NUM'].apply(lambda x: cov_names(x)).fillna('NA'))

#df_name[df_name['S_NUM'].str.len() < 8]

#df_name[df_name['S_NUM'].str.contains('COVC') == False]

### **Clean Age**

#df_name[df_name['AGE_UNIT'] == 'Days']

df_months = df_name[df_name['AGE_UNIT'] == 'Months']

#df_months.head()

df_mn2yrs = df_months.assign(AGE=df_months['AGE'].map(lambda x: round(x / 12, 1))).replace('Months', 'Years')

#df_mn2yrs.head()

#df_name[df_name['AGE_UNIT'] == 'Days']#.AGE_UNIT.unique()

df_not_months = df_name[df_name['AGE_UNIT'] != 'Months']

df_not_months.shape

df_mn2yrs.shape

df_years = pd.concat([df_not_months, df_mn2yrs])

df_years.shape

df_years['AGE_UNIT'].unique()

#df_years[df_years['AGE_UNIT'] == 'NO PATIENT DATASHEET']

#df_years[df_years['AGE_UNIT'].str.contains('M', 'F') == True]

#df_years[df_years['AGE_UNIT'] == 'Nil']

df_years_drop_age_unit = df_years.drop('AGE_UNIT', axis=1)

df_years_rencol = df_years_drop_age_unit.rename(columns={'AGE': 'AGE_YRS'})

#df_years_rencol.head()

df_years_rencol['AGE_YRS'].unique()

# [x for x in df_years_rencol['AGE_YRS'].unique() if isinstance(x, str)]

to_rep = ['None indicated',
 'Nil',
 'Not indicated',
 'None',
 'nil',
 'none indicated',
 'not_indicated',
 'Not Indicated',
 'not Indicated',
 'None Indicated',
 'NO PATIENT DATASHEET',
 'ad', 'AD'
]

df_years_repnan2float = df_years_rencol.assign(AGE_YRS=df_years_rencol['AGE_YRS'].replace(to_rep, np.nan).astype(float).fillna(np.nan))

#df_years_repnan2float

### **Clean Gender**

df_years_repnan2float['GEND'].unique()

#df_years_repnan2float[df_years_repnan2float['GEND'].isna()].sort_values('S_NUM')

#df_years_repnan2float[df_years_repnan2float['GEND'] == 'Nil']

COVEs:
- COVC13844
- COVC13845
- COVC13846
- COVC13847
- COVC13848
- COVC13849
- COVC13850
- COVC13851
- COVC13852
- COVC13853
- COVC16256
- COVC16257

df_years_repgender = df_years_repnan2float.assign(GEND=df_years_repnan2float['GEND'].str.strip().replace(['f', 'Female', 'F`', 'T', 'FF'], 'F')
                                                 .replace(['Male', 'MALE ', 'MALE', 'm', 'MM'], 'M'))

df_years_repgender['GEND'].unique()

df_years_repgender2 = df_years_repgender.assign(GEND=df_years_repgender['GEND'].replace('Nil', pd.NA).fillna('NA'))

# df_years_repgender2['GEND'].unique()

### **Clean Occupation**

# df_years_repgender2['OCCU'].unique()

# df_years_repgender2['OCCU'] = df_years_repgender2['OCCU'].replace(['Nil','nil'], pd.NA)

df_years_repgender2['OCCU'].str.lower().str.strip().nunique()

occu = (df_years_repgender2['OCCU'].str.strip().
     apply(lambda x: str(x).replace(str(x), 'NA') if ('nil' == str(x).lower()) else x).
     replace('nan', 'NA').fillna('NA'))
occu.nunique()

df_occu = df_years_repgender2.assign(OCCU=occu)

### **Clean Nationality**

df_occu['NAT'].unique()

df_years_repnat = df_occu.assign(NAT=df_occu['NAT'].replace('Nil', 'Unknown')
                                            .replace(['United States', 'Indian'
                                                      , 'Congo, Democratic Republic of the'
                                                     , 'Congo, Republic of the'] 
                                                     , ['United States of America (USA)','India'
                                                        , 'Democratic Republic of the Congo',
                                                       'Republic of the Congo'])
                                            .apply(lambda x: str(x).replace(x, 'Kenya') if ('enya' in str(x).lower()) else x).fillna('NA'))

# df_years_repnat['NAT'].unique()

### **Clean County of Res**

#df_years_repnat.head()

counties = (df_years_repnat['COUNT_RES']
                   .replace(['Nil', 'nan', '', 'Nan', 'Muranga ', 'Murang`a ', 'Niarobi']
                            , [pd.NA, pd.NA, pd.NA, pd.NA, "Murang'a", "Murang'a", 'Nairobi'])
                   .apply(lambda x: str(x).strip().capitalize() if (isinstance(x, str)) else x)).fillna('NA')

df_years_county = (df_years_repnat.assign(COUNT_RES=counties))#.assign(S_COUNT_RES=s_counties)

# df_years_county['COUNT_RES'].unique()

### **Clean Subcounty of Res**

df_years_repnat['S_COUNT_RES'].str.lower().nunique()

len((df_years_repnat['S_COUNT_RES'].fillna('NA')
     .replace(['Nil', 'Nil ', 'Muranga ', 'Murang`a ']
                            , [pd.NA, pd.NA, "Murang'a", "Murang'a"])
                   .apply(lambda x: str(x).strip().capitalize())
     .unique()))

s_counties = (df_years_repnat['S_COUNT_RES'].str.strip().str.capitalize().
    replace(['Nil', 'Muranga', 'Murang`a'],
    ['NA', "Murang'a", "Murang'a"]).
    apply(lambda x: str(x).replace(str(x), 'NA') if ('indicate' in str(x)) else x))

# sorted([s for s in s_counties.unique()])

df_years_scounty = df_years_county.assign(S_COUNT_RES=s_counties)

### **Clean Travel History**

#df_years_scounty.head()

df_years_scounty['TRAV_HIST'].unique()

df_travl = df_years_scounty.assign(TRAV_HIST=df_years_scounty['TRAV_HIST'].replace([pd.NA, 'N', 'no', 'NO', 'Nil', 'NONE ', 'NONE'], 'No')
           .replace(['YES', 'Y', 'y', 'yes'], 'Yes'))


df_travl['TRAV_FROM'].unique()

df_travl2 = (df_travl.assign(TRAV_FROM=df_travl['TRAV_FROM'].str.strip().replace(['nan','NIL', 'Nil', 'Not indicated on form '
                  , 'NOT APPLICABLE ', 'N', 'N', 'No', 'not_indicated', 'Not applicable', 'Not indicated'],pd.NA)
                    .replace(['Mombasa (stayed in mtwapa for 4 days)', 'KINANGO MSAMBWENI', 'Nan'
                              , 'Mombasa (Stayed in Mtwapa for 4 days)'], 
                    ['Mtwapa', 'Kinangop;Msambweni', pd.NA, 'Mtwapa'])
                    .apply(lambda x: str(x).replace('/', ';')
                    .replace(' and ', ';')
                    .replace('-', ';')
                    .replace(',', ';')
                    .replace(' ; ', ';')
                    .replace('; ', ';')
                    .strip().capitalize() if (isinstance(x, str)) else x)).fillna('NA'))

# df_travl2['TRAV_FROM'].unique()



df_travl2.head(1)

### **Clean Quarentine Place**

df_travl2['QUAR_PLACE'].str.lower().str.strip().nunique()

quar_pl = (df_travl2['QUAR_PLACE'].str.strip().
     apply(lambda x: str(x).replace(str(x), 'NA') if ('nil' == str(x).lower()) else x).
     apply(lambda x: str(x).replace(str(x), 'NA') if ('indicate' in str(x).lower()) else x).
     apply(lambda x: str(x).replace(str(x), 'NA') if ('applic' in str(x).lower()) else x).
     apply(lambda x: str(x).replace(str(x), 'NA') if ('n/a' in str(x).lower()) else x).      
     fillna('NA'))
quar_pl.nunique()

df_quar = df_travl2.assign(QUAR_PLACE=quar_pl)

### **Clean Symptoms**

Symptoms legend
- GW: General (body) weakness/fatigue/malaise
- FC: Fever/Chills
- CO: Cough
- ST: Sore throat
- RN: Runny nose
- SB: Shortness of breath/Difficulty in breathing
- D: Diarrhoea
- NV: Nausea/Vomiting
- H: Headache
- IC: Irritability/Confusion
- P: Pain
- LS: Loss of smell
- LT: Loss of taste/appetite
- P-M: Muscular pain
- P-A: Abdominal pain
- P-B: Back pain
- P-C: Chest pain
- P-J: Joint pain
- BA: Body aches
- SZ: Sneezing
- HP: Hypertension
- TN: Tonsillitis
- DZ: Dizziness
- NS: Night sweats
- SG: Swollen gland

**Symptoms table**

abbr = ['GW','FC','CO','ST','RN','SB','D','NV','H',
        'IC','P','LS','LT','P-M','P-A','P-B','P-C',
        'P-J','BA','SZ','HP','TN','DZ','NS', 'SG']
symp = ['General (body) weakness/fatigue/malaise','Fever/Chills',
        'Cough','Sore throat','Runny nose','Shortness of breath/Difficulty in breathing',
        'Diarrhoea','Nausea/Vomiting','Headache','Irritability/Confusion','Pain','Loss of smell',
        'Loss of taste/appetite','Muscular pain','Abdominal pain','Back pain','Chest pain',
        'Joint pain','Body aches','Sneezing','Hypertension','Tonsillitis','Dizziness','Night sweats', 'Swollen glands']
symp_header = pd.MultiIndex.from_arrays([['SYMP_CODE', 'SC_MEANING'], 
             ['Symptom code', 'Code meaning']], names=['abbr', 'desc'])
symptoms = pd.DataFrame({'a': abbr, 'b': symp})
symptoms.columns=symp_header
symptoms.head()

df_quar['SYMPS'] = (df_quar['SYMPS'].apply(lambda x: str(x).replace(str(x), 'NA') if ('symtom' in str(x)) else str(x).replace('F/C', 'FC'))
                      .replace(['[=]','nan'], 'NA'))
pd.Series(df_quar['SYMPS']).to_excel(f"{glob.glob(f'{parent_dir}/**/Outputs', recursive=True)[0]}/COVID19-results-merged-symptoms.xlsx", index=False)
df_symp = pd.read_excel(f"{glob.glob(f'{parent_dir}/**/Outputs', recursive=True)[0]}/COVID19-results-merged-symptoms_curation.xlsx")
symp_dict = {str(k): str(v) for k,v in zip(df_symp['STRING'], df_symp['REPLACE'])}

symp_cln = (df_quar['SYMPS'].replace((symp_dict)).replace(['nan'], ['NA']).str.strip().
replace(['; ', ',', '.'], ';').replace(';;', ';').apply(lambda x: ';'.join(sorted(x.split(';')))))

df_symps = df_quar.assign(SYMPS=symp_cln.fillna('NA'))

### **Clean Sample Type**

df_symps.SAMP_TYPE.unique()
df_samp_ty = df_symps.assign(SAMP_TYPE=df_symps['SAMP_TYPE'].replace(['NP and OP', 'NP-OP +AB2:AL2Swab',
       'NP-OP Swab', 'nP Swab', 'NP OP Swab', 'NP&OP'], ['NP & OP Swab', 'NP & OP Swab', 'NP & OP Swab', 'NP Swab', 'NP & OP Swab', 'NP & OP Swab']).fillna('NA'))

### **Clean Dates**

df_dt1 = df_samp_ty.assign(DT_SAM_COLL=df_samp_ty['DT_SAM_COLL'].apply(lambda x: my_date_parser(x)))

df_dt2 = df_dt1.assign(DT_SAM_RECEP=df_dt1['DT_SAM_RECEP'].apply(lambda x: my_date_parser(x)))

df_dts_cln = df_dt2.assign(DT_CONF=df_dt2['DT_CONF'].apply(lambda x: my_date_parser(x)))

#df_dts_cln.head()

### **Clean Results**

df_dts_cln['RESULT'].unique()

df_dts_cln['RESULT'].map(lambda x: x.strip().lower() if (isinstance(x, str)) else x).unique()

df_dts_cln['RESULT'] = df_dts_cln['RESULT'].map(lambda x: cln_results(x)).fillna('NA').replace('na', 'NA')

df_dts_cln['RESULT'].unique()

df_dts_cln[df_dts_cln.duplicated(keep=False) == True].sort_values('S_NUM')

### **Remove/QC Duplicates**

df_cln_srt = df_dts_cln.sort_values(['S_NUM', 'DT_CONF'])

#df_cln_srt[df_cln_srt.duplicated(['S_NUM', 'DT_CONF'], keep=False) == True]

#df_cln_srt[df_cln_srt.duplicated(keep=False) == True]

df_dedud0 = df_cln_srt.drop_duplicates(['S_NUM'], keep='last')#['COUNT_RES'].fillna('Unknown', inplace=True)
df_dedud = df_dedud0.assign(CASE_ID=df_dedud0.CASE_ID.apply(lambda x: str(x)))

mask1 = df_dedud.duplicated(['CASE_ID'], keep=False)
mask2 = df_dedud.CASE_ID != 'NA'
df_dedud.sort_values(['DT_CONF','CASE_ID'])[mask1 == True & mask2].tail(10)

df_dedud[df_dedud.duplicated(['S_NUM'], keep=False) == True]

### **QC Dates**

sr_date_diff = df_dedud['DT_CONF'] - df_dedud['DT_SAM_COLL']

mask3 = sr_date_diff >= '60D'
mask4 = sr_date_diff != pd.NaT
mask5 = df_dedud['DT_SAM_COLL'] > df_dedud['DT_CONF']
mask6 = df_dedud['DT_SAM_COLL'] > dt
mask7 = df_dedud['DT_CONF'] > dt
# mask3 = df_dedud['DT_CONF']

df_dedud[mask5 | mask6 | mask7 == True]

### **Export Cleaned Data**

header_des1 = pd.MultiIndex.from_arrays([['CASE_ID', 'S_NUM', 'AGE_YRS', 
               'GEND', 'OCCU', 'NAT', 'COUNT_RES', 'S_COUNT_RES', 'TRAV_HIST', 
               'TRAV_FROM', 'QUAR_PLACE', 'SYMPS', 'SAMP_TYPE', 'DT_SAM_COLL', 
               'DT_SAM_RECEP', 'RESULT', 'DT_CONF', 'VACC_STATUS', 'VACC_DOS'], 
               ['Source sample ID', 'Unique lab ID', 'Age in years', 'Gender', 
               'Occupation', 'Nationality', 'County of residence', 'Sub-county of residence', 
               'Travel history', 'Place travelled', 'Place of quarantine', 
               'Symptoms', 'Sample type', 'Date of sample collection', 
               'Date of sample lab reception', 'RT-PCR results', 'Date of RT-PCR', 
               'Vaccination status', 'Vaccine dosage']], names=('abbr', 'desc'))

df_dedud.columns = header_des1

df_dedud.to_excel(f"{glob.glob(f'{parent_dir}/**/Outputs', recursive=True)[0]}/COVID19-results-merged-cln.xlsx", index=True)

df_dedud[df_dedud['RESULT'] == 'Positive'].to_excel(f"{glob.glob(f'{parent_dir}/**/Outputs', recursive=True)[0]}/COVID19-results-merged-cln-pos.xlsx", header=header_des1, index=True)

#df_dedud[df_dedud['S_NUM'].str.contains('COVC') == False]

#df_dedud[df_dedud['S_NUM'].str.contains('COVC') == False]

df_symps11-09-2020

COVC9774
COVC9775
COVE20
COVE21
COVC9776
COVC9777


### **ID Misssing Data**

start = 0
start2 = 0
df_missing = pd.DataFrame(columns=cols_rename)
df_missing2 = pd.DataFrame(columns=cols_rename)
for num in df_dedud[('S_NUM','Unique lab ID') ].sort_values():
#     curr_row = df_dts_cln[df_dts_cln['S_NUM']]
    try:
        curr = int(num.lstrip('COVC').lstrip('0'))
        curr2 = int(num.lstrip('COVM').lstrip('0'))
        
        if curr - start > 1:
            start_id = f'COVC{start}'
            start_id2 = f'COVM{start}'
            cov_id = f'COVC{curr}'
            cov_id2 = f'COVM{curr2}'
            df_missing = pd.concat([df_missing, df_dts_cln[df_dts_cln['S_NUM'] == cov_names(start_id)]])
            df_missing2 = pd.concat([df_missing, df_dts_cln[df_dts_cln['S_NUM'] == start_id2]])
            df_missing = pd.concat([df_missing, df_dts_cln[df_dts_cln['S_NUM'] == cov_names(cov_id)]])
            df_missing2 = pd.concat([df_missing, df_dts_cln[df_dts_cln['S_NUM'] == cov_id2]])
        start = curr
        start2 = curr2
    except ValueError:
#         print(f'ValueError: {num}')
        pass

#df_missing.head()

(df_missing[['S_NUM', 'DT_SAM_COLL', 'DT_SAM_RECEP', 'DT_CONF']]
 .sort_values(['S_NUM', 'DT_CONF'])
 .drop_duplicates('S_NUM', keep='last')
 .to_excel(f"{glob.glob(f'{parent_dir}/**/Outputs', recursive=True)[0]}/COVID19-results-merged-missing.xlsx", index=False))

### **ID Missing Results**

(df_dts_cln[df_dts_cln['RESULT'].isin(['Positive', 'Negative']) == 
            False][['S_NUM', 'DT_SAM_COLL', 'DT_SAM_RECEP', 'DT_CONF', 'RESULT']]
.to_excel(f"{glob.glob(f'{parent_dir}/**/Outputs', recursive=True)[0]}/COVID19-results-merged-results-missing.xlsx", index=False))

# plt.style.available

### **Metadata + Cts**

df_Cts = pd.read_excel(f"{glob.glob(f'{parent_dir}/**/Outputs', recursive=True)[0]}/runsheet-cts.xlsx")[['Sample_Name', 'Well_Position', 'Ct_Mean', 'Dt_Run']]

header_des2 = pd.MultiIndex.from_arrays([['S_NUM', 'RTPCR_POS', 'AVG_Ct', 'DT_RUN'], 
               ['Unique lab ID', 'RT-PCR Well position', 'Detection genes average Ct', 'RT-PCR run date']])

df_Cts.columns = header_des2

df_metCts = df_dedud.set_index(('S_NUM', 'Unique lab ID')).merge(df_Cts.set_index(('S_NUM', 'Unique lab ID')), how='left', left_index=True, right_index=True).reset_index().fillna('NA')#.rename(columns={'index': 'S_NUM', 'Ct_Mean': 'AVG_Ct'})

writer = pd.ExcelWriter(f"{glob.glob(f'{parent_dir}/**/Outputs', recursive=True)[0]}/COVID19-resultsCts-merged-cln.xlsx")
df_metCts.to_excel(writer, sheet_name='Metadata', index=True, na_rep='NA', float_format='%.1f')
symptoms.to_excel(writer, sheet_name='symptomsDictionary', index=True, na_rep='NA', float_format='%.1f')
writer.close()

df_metCts.shape

df_metCts[df_metCts['RESULT'] == 'Positive'].to_excel(f"{glob.glob(f'{parent_dir}/**/Outputs', recursive=True)[0]}/COVID19-resultsCts-merged-cln-pos.xlsx", index=True)

#### **Confirm updated file in the following cell**

df_sequenced_samples = pd.read_excel(f"{glob.glob(f'{parent_dir}/**/Outputs', recursive=True)[0]}/all-sequenced-samples-IDs_20-04-2022.xlsx")
seqd_list = list(df_sequenced_samples['SAMPLE'])

df_dedud_pos = df_metCts[df_metCts['RESULT'] == 'Positive']
df_dedud_pos_unseq = df_dedud_pos[df_dedud_pos['S_NUM'].isin(seqd_list) == False]
df_dedud_pos_unseq.to_excel(f"{glob.glob(f'{parent_dir}/**/Outputs', recursive=True)[0]}/all-unsequenced-pos-samples_{dt}.xlsx", index=True)

#df_dedud_pos_unseq

df_dedud_pos = df_metCts[df_metCts['RESULT'] == 'Positive']
df_dedud_pos_seq = df_dedud_pos[df_dedud_pos['S_NUM'].isin(seqd_list) == True]
df_dedud_pos_seq.to_excel(f"{glob.glob(f'{parent_dir}/**/Outputs', recursive=True)[0]}/all-sequenced-pos-samples_{dt}.xlsx", index=True)

### Positives 2021 - Unsequenced

df_dedud_pos_unseq_21 = df_dedud_pos_unseq[df_dedud_pos_unseq['DT_CONF'] >= pd.to_datetime('2021-01-01 00:00:00')]
df_dedud_pos_unseq_21.to_excel(f"{glob.glob(f'{parent_dir}/**/Outputs', recursive=True)[0]}/all-unsequenced-pos-samples-2021_{dt}.xlsx", index=True)
