==> 03.aggregate-analysis-reports-nf-viralrecon.ipynb <==
## **Notebook for aggregating sequencing reports**

### **Packages**

import pandas as pd
import seaborn as sns
import numpy as np
import glob, os, re
import matplotlib.pyplot as plt
from datetime import  datetime
from ipywidgets import widgets, interactive
from pandas import ExcelWriter

dt = datetime.today().strftime(format='%d-%m-%Y')

### **Directories & Files**
Uniqueness in directory and file names is assumed for all analyses

The organisation of the `run_dir`: The directory name MUST be unique and reside anywhere inside `sars_dir` directory
 
| Directory name | File name | File source-tool | File description |
| :-------------- | :--------- | :---------------- | :------|
|qcs *(QCs)*|`*.tsv`|Quast|Transposed report|
|nxt *(nextclade)*|`nxt.tsv`|Nextclade|Renamed `nextclade.tsv` output|
|snpEff|`*sneff.vcf`| snpEff|Per-sample unzipped `snpEff` output|
|var *(variants)*|`k-per-gene_variant_anns.tsv`|Script: `abstract_snpeff_ann_output.py`|Aggregation of individual `snpEff.vcf` outputs by abstracting gene-mutations|
|png *(pangolin)*|`png.csv`|Pangolin|Renamed `Pangolin result.csv` output|
|dpt *(coverage depths)*|`amplicon, genome`|Mosdepth|Per-sample amplicon/genome depths<br> ***Cols**:chrom, start, end, region, coverage, sample*|
|plt *(plot)*|`*.pdf`|Quast|Heatmap plot for median read depth coverage|

### **Preliminary variables**

sars_dir = 'bioinformatics/github/covid/data/SarsGenomics/SARS-CoV-2' #  name of root directory for all SARS-associated work
home_dir = os.getenv('HOME') #  get OS home directory
parent_dir = glob.glob(f'{home_dir}/{sars_dir}', recursive=True)[0]

### **Metadata**

# import raw metadata file
df_rmd_cln = pd.read_excel(glob.glob(f'{parent_dir}/**/Outputs/COVID19-resultsCts-merged-cln.xlsx', recursive=True)[0]).rename(columns={'S_NUM': 'sample_name'}).drop('abbr', axis=1)

### **Functions**

# define a func to replace spaces in the header names
def tidy_header(df):
    df.columns = [col.replace(' ', '_') for col in df.columns]
    return df

# define func to retrieve particular columns from a df (spaces in col names must be replaced with _ in the input col_list)
def get_cols(df, col_list):
    new_df = tidy_header(df)
    return new_df[col_list]

# define a function to replace from a dictionary ('key is what is to be replaced': 'value is the replacement')
def replace(string, substitutions):
    substrings = sorted(substitutions, key=len, reverse=True)
    regex = re.compile('|'.join(map(re.escape, substrings)))
    return regex.sub(lambda match: substitutions[match.group(0)], string)

# define function to merge variants and nextclade data
def merge_varNxt(df_var_cln, df_nxt_cln):
    return (df_var_cln.set_index('sample_name').merge(df_nxt_cln
          .set_index('seqName'), how='outer', left_index=True, right_index=True)
                 .reset_index().rename(columns={'index': 'sample_name'}))

# define function to merge pangolin and variants-nextclade data
def merge_pngVxt(df_png_cln, df_varNxt):
    return (df_png_cln.set_index('Sequence_name').merge(df_varNxt
        .set_index('sample_name'), how='outer', left_index=True, right_index=True)
            .reset_index().rename(columns={'index': 'Sequence_name'}))


# define function to merge metadata with cts data
def merge_rmdCts(df_rmd_cln, df_cts_cln):
    return (df_rmd_cln.set_index('S_NUM').merge(df_cts_cln.set_index('Sample_Name'), how='outer', left_index=True, right_index=True)
            .reset_index().rename(columns={'index': 'S_NUM'}))


# define function to merge metadata and seq data
def merge_vnpPmd(df_pngVxt, df_rmdCts):
    return (df_pngVxt.set_index('S_NUM')
            .merge(df_rmdCts.set_index('S_NUM'), how='left', left_index=True, right_index=True)
                 .reset_index().rename(columns={'index': 'S_NUM'}))

# define a function to retrieve MoC and all mutations for the s-gene
def get_mut_of_concern(ann_file_name, moc_list):

    def intersection(x, y):
        return list(set(x) & set(y))

    moc_list = moc_list
#     file_name = 'k-per-gene_variant_anns.tsv'
    df = ann_file_name[['sample_name','S']]

#     df = pd.read_table(f'{base_dir}/{file_name}')[['sample_name','S']]
    mutations = []
    moc = []
    sample_id = []
    for row in df.itertuples():
        if isinstance(row.S, str):
            sgene = row.S
        else: 
            sgene = str(row.S)
        substitutions = sgene.replace(' ', '').split(',')[1:-1]
        if len(moc_list) >= len(intersection(moc_list, substitutions)) > 0:
            intsct = intersection(moc_list, substitutions)
            sample_name = row.sample_name
            mutations.append(str(substitutions).replace("[", "").replace("]", "").replace("'", ""))
            moc.append(str(intsct).replace("[", "").replace("]", "").replace("'", ""))
            sample_id.append(sample_name)
        else: pass 
    df = pd.DataFrame({'Sample_ID': sample_id, 'Mut_of_Concern_(S)': moc, 'All_Mutations_(S)': mutations})
    df_fnl = df.assign(Sample_ID = df['Sample_ID'].apply(lambda x: x.split('_')[0].split('.')[0]))
    return df_fnl

def replace_with_who_lin(x):
    if x == 'B.1.1.7':
        return x.replace(x, 'B.1.1.7(Alpha)')
    elif x == 'B.1.617.2':
        return x.replace(x, 'B.1.617.2(Delta)')
    elif x == 'B.1.351':
        return x.replace(x, 'B.1.351(Beta)')
    elif x == 'B.1.525':
        return x.replace(x, 'B.1.525(Eta)')
    elif 'AY' in str(x):
        return str(x).replace(str(x), str(x)+'(Delta)')
    elif x == 'B.1.1.529':
        return x.replace(x, 'B.1.1.529(Omicron)')
    elif 'BA' in str(x):
        return str(x).replace(str(x), str(x)+'(Omicron)')
    return x   

### **Variables**

Reassign accordingly...

pipeline = 'nf-viralrecon-v2.5' #  name and version
seq_name = 'seq52' #  seq*
tech = 'NextSeq' #  NextSeq/MiSeq/MinION
seq_dt = '23/08/2022' #  DD/MM/YYYY
lib_prep = 'COVIDSeq' #  NEBNext/NEBNext_FS/COVIDSeq/Nextera_XT
primer_set = 'ARTIC_V3' #  ARTIC_V3/ARTIC_V4
identifier = 'ILL_seq52'#  used in naming file outputs
run_dir = '2022-08-23_run52_nextseq' #  name of the run directory containing viralcon pipeline output as implemented by Kibet

### **Sequencing sheet**

# import sequencing cheat sheet
df_seq_sh = (pd.read_excel(glob.glob(f'{parent_dir}/**/SeqSampleSheets/index_cheat_sheet_{seq_name}.xlsx', recursive=True)[0], usecols=['indexing', 'idx_plt_pos']).
             rename(columns={'indexing': 'sample_name'}))
df_seq_sh_fnl = df_seq_sh[df_seq_sh['sample_name'].str.contains('COV')]
df_seq_sh_fnl.head(2)

### **QCstats**

# #import the collated file for all the multiqc output
# nxt = pd.read_table(glob.glob(f'{parent_dir}/**/{run_dir}/nxt/nxt.tsv', recursive=True)[0])[['seqName', 'totalMissing']]
# nxt.seqName = nxt.seqName.apply(lambda x: x.replace(' MN908947.3', ''))
# nxt.totalMissing = nxt.totalMissing.apply(lambda x: 100 - round((x/29903)*100,1))
# df_qcs_trans_fnl = nxt.rename(columns={'seqName': 'sample_name', 'totalMissing': 'Genome fraction (%)'})

# import the collated file for all the multiqc output
df_qcs_trans_cols = ['sample_name', 'Genome fraction (%)']# , 'Assembly'
df_qcs_trans = pd.read_table(glob.glob(f'{parent_dir}/**/{run_dir}/qcs/transposed_report.tsv', recursive=True)[0])
df_qcs_trans2 = df_qcs_trans.assign(sample_name = df_qcs_trans['Assembly'].apply(lambda x: x.split('_')[0].split('.')[0]))[df_qcs_trans_cols].rename(columns={'Genome fraction (%)': 'genome_cov'})
df_qcs_trans_fnl = df_qcs_trans2.assign(genome_cov=df_qcs_trans2['genome_cov'].replace('-', np.nan).apply(lambda x: round(float(x),1) if float(x) else np.nan)).rename(columns={'genome_cov': 'Genome fraction (%)'})
# df_qcs_trans_fnl['Seq id'] = run_dir
# df_qcs_fnl['Analysis type'] = 'Qcs'

df_qcs_trans_fnl.head(10)

# import the collated file for all the multiqc output
if 'Seq' in tech or 'Illumina' in tech:
    sample_name = 'Sample'
    df_qcs_cols = ['sample_name', '# Input reads', '# Trimmed reads (fastp)',
       '% Mapped reads', '# Mapped reads', '# Trimmed reads (iVar)', 
       'Coverage median', '% Coverage > 1x', '% Coverage > 10x',
       'Pangolin lineage', 'Nextclade clade']#  , 'Sample'
else:
    df_qcs_cols = ['Sample', '# Mapped reads', 'Coverage median', '% Coverage > 1x', '% Coverage > 10x',
       'Pangolin lineage', 'Nextclade clade']
    sample_name = 'Sample'

df_qcs = pd.read_csv(glob.glob(f'{parent_dir}/**/{run_dir}/qcs/summary_variants_metrics_mqc.csv', recursive=True)[0], sep=',')
df_qcs_fnl = df_qcs.assign(sample_name = df_qcs[sample_name].apply(lambda x: x.split('_')[0]))[df_qcs_cols]
if tech == 'MinION':
    df_qcs_fnl[['# Input reads', '# Trimmed reads (fastp)', '% Mapped reads', '# Trimmed reads (iVar)']] = [np.nan, np.nan, np.nan, np.nan]
    df_qcs_fnl = df_qcs_fnl[['Sample', '# Input reads', '# Trimmed reads (fastp)',
       '% Mapped reads', '# Mapped reads', '# Trimmed reads (iVar)', 
       'Coverage median', '% Coverage > 1x', '% Coverage > 10x',
       'Pangolin lineage', 'Nextclade clade']].rename(columns={'Sample': 'sample_name'})
df_qcs_fnl['Seq id'] = run_dir
# df_qcs_fnl['Analysis type'] = 'Qcs'

df_qcs_fnl.head(1)

df_seq_qcs = df_seq_sh_fnl.merge(df_qcs_trans_fnl, left_on='sample_name', right_on='sample_name', how='left')#.sort_values('Genome fraction (%)', ascending=False)

df_seq_qcs.head(1)

qcStat = df_seq_qcs.merge(df_qcs_fnl, left_on='sample_name', right_on='sample_name', how='left').sort_values('Genome fraction (%)', ascending=False)

df_seq_meta = df_seq_qcs.merge(df_rmd_cln, left_on='sample_name', right_on='sample_name', how='left').sort_values('Genome fraction (%)', ascending=False)
# df_seq_meta.head()

qcStat.head(1)

### **Nextclade data**

# import Nextclade clade data
df_nxt = pd.read_table(glob.glob(f'{parent_dir}/**/{run_dir}/nxt/nxt.tsv', recursive=True)[0])

# sep = '.' if 'Seq' in tech else '_'

# coverage based on totalMissing col in nexclade output
coverage = round(100 - (df_nxt['totalMissing'] / 29903) * 100, 1)
df_nxt_cln1 = df_nxt.assign(seqName = df_nxt['seqName'].apply(lambda x: x.split('_')[0])).rename(columns={'seqName': 'sample_name'})
df_nxt_cln2 = df_nxt_cln1.assign(coverage = coverage)
df_nxt_fnl = df_nxt_cln1
df_nxt_fnl['Seq id'] = run_dir
# df_nxt_fnl['Analysis type'] = 'Nxt'

nextclade = df_seq_qcs.merge(df_nxt_fnl, left_on='sample_name', right_on='sample_name', how='left').sort_values('Genome fraction (%)', ascending=False)
nextclade.head(1)

### **Variants data**

# import the collated file for all the snpEff outputs
df_var = pd.read_csv(f"{glob.glob(f'{parent_dir}/**/{run_dir}/var', recursive=True)[0]}/k-per-gene_variant_anns.tsv", sep='\t')
df_var_fnl = df_var.assign(sample_name = df_var['sample_name'].apply(lambda x: x.split('_')[0].split('.')[0]))
# df_var_fnl['Seq id'] = df_var_fnl.sample_name.
# df_var_fnl['Analysis type'] = 'Var'

df_var_fnl.head(1)

# MOH mutations of concern (spike only)
moc_list = ['N501Y', 'E484K', 'K417N', 'P681R', 'E484Q', 'L452R', 'A570D', 'D80A', 'Q677H', 'F888L', 'L5F', 'D253G', 'E154K', 'A67V',
           'D614G', 'D796Y', 'E484A', 'G339D', 'G446S', 'G496S', 'H655Y', 'L212I', 'L981F', 'N440K', 'N679K', 'N764K', 'N856K',
           'N969K', 'P681H', 'Q493R', 'Q498R', 'Q954H', 'S371L', 'S373P', 'S375F', 'S477N', 'T478K', 'T547K', 'T95I', 'Y145D',
           'Y505H',  'del142-144', 'del211', 'del69-70', 'ins214EPE']
file_name = df_var_fnl
df_moc = get_mut_of_concern(file_name, moc_list)

iVar_snpE = df_seq_qcs.merge(df_var_fnl, left_on='sample_name', right_on='sample_name', how='left').sort_values('Genome fraction (%)', ascending=False)
iVar_snpE.head(1)

### **Pangolin data**

# import Pangolin lineage data
# df_png = pd.read_csv(glob.glob(f'{parent_dir}/**/{run_dir}/png/png.csv', recursive=True)[0])
df_png_web = pd.read_csv(glob.glob(f'{parent_dir}/**/{run_dir}/png/png.csv', recursive=True)[0]).rename(columns={'Sequence name': 'sample_name'})

months = {'January': 'Jan', 'February': 'Feb', 'March': 'Mar',
         'April': 'Apr', 'June': 'Jun', 'July': 'Jul', 'August': 'Aug',
          'September': 'Sep', 'October': 'Oct', 'November': 'Nov', 'December': 'Dec'}
# retrieve cols Sequence_name and Lineage (func get_cols replaces col names spaces with _)
# cols = ['taxon', 'lineage', 'scorpio_call']#, 'Most_common_countries']
# df_png_cln = get_cols(tidy_header(df_png), cols)
# df_png_fnl = (df_png_cln.assign(taxon = df_png_cln['taxon']
#             .apply(lambda x: '_'.join(x.split('_')[1:2]) if (len(x
#             .split('_')) > 2) else x.split('_')[0]))).rename(columns={'taxon': 'sample_name'})
df_png_fnl_web = (df_png_web.assign(sample_name = df_png_web['sample_name']
                .apply(lambda x: x.split('_')[0])))
# df_png_fnl = df_png_cln1.assign(Date_range=df_png_cln1['Date_range'].apply(lambda x: replace(x, months) if (isinstance(x, str)) else x))
df_png_fnl_web['Seq id'] = run_dir
# df_png_fnl_web['Analysis type'] = 'Png'

pangolin = df_seq_qcs.merge(df_png_fnl_web, left_on='sample_name', right_on='sample_name', how='left').sort_values('Genome fraction (%)', ascending=False)
pangolin.head(1)

### **Summary: combining data**

#### *Merge summary data*

pango_summ = pangolin[['sample_name', 'Lineage', 'Scorpio call']]
next_summ = nextclade[['sample_name', 'clade']]
var_s = iVar_snpE[['sample_name', 'S']]
var_summ = var_s.assign(spike_mut = var_s['S'].map(lambda x: int(len(x.split(',')))  if isinstance(x, str) else 0))
pango_next = pango_summ.merge(next_summ, left_on='sample_name', right_on='sample_name', how='outer')
pn_var = pango_next.merge(var_summ, left_on='sample_name', right_on='sample_name', how='outer')
meta_summ = df_seq_meta[['CASE_ID', 'sample_name']]

df_comb = df_seq_qcs.merge(pn_var, left_on='sample_name', right_on='sample_name', how='left')
summary = (meta_summ.merge(df_comb, left_on='sample_name', right_on='sample_name', how='right').sort_values('Genome fraction (%)', ascending=False)
          .rename(columns={'CASE_ID': 'Case id', 'sample_name': 'Unique lab id', 'plt_pos': 'Seq plate pos', 'clade': 'Clade', 'S': 'Spike mutations (snpEff)', 'spike_mut':'Spike mutation count'}))

summary[['Seq number', 'Seq platform', 'Seq date', 'Library kit', 'Primer set', 'Analysis pipeline']] = [seq_name, tech, seq_dt, lib_prep, primer_set, pipeline]

summary.head(1)

### **Symptoms table**

abbr = ['GW','FC','CO','ST','RN','SB','D','NV','H',
        'IC','P','LS','LT','P-M','P-A','P-B','P-C',
        'P-J','BA','SZ','HP','TN','DZ','NS', 'SB']
symp = ['General (body) weakness/fatigue/malaise','Fever/Chills',
        'Cough','Sore throat','Runny nose','Shortness of breath/Difficulty in breathing',
        'Diarrhoea','Nausea/Vomiting','Headache','Irritability/Confusion','Pain','Loss of smell',
        'Loss of taste/appetite','Muscular pain','Abdominal pain','Back pain','Chest pain',
        'Joint pain','Body aches','Sneezing','Hypertension','Tonsillitis','Dizziness','Night sweats', 'Swollen glands']
symp_header = pd.MultiIndex.from_arrays([['SYMP_CODE', 'SC_DESC'], 
             ['Symptom code', 'Code description']], names=['abbr', 'desc'])
symptoms = pd.DataFrame({'a': abbr, 'b': symp})
symptoms.columns=symp_header
symptoms.head()

### **Generate reports**

writer = pd.ExcelWriter(f"{glob.glob(f'{parent_dir}/**/SeqReports', recursive=True)[0]}/{run_dir}.Analysis.QCstats.xlsx")
datasheets = [(qcStat, 'QCstats'), (pangolin, 'pangolinAnalysis'), (nextclade, 'nextcladeAnalysis'), (iVar_snpE, 'snpEffAnnotation'), (df_seq_meta, 'metaData'), (summary, 'summaryReport')]
for datasheet in datasheets:
    samp_id = 'sample_name' if datasheet[1] != 'summaryReport' else 'Unique lab id'
    (datasheet[0].sort_values([samp_id, 'Genome fraction (%)'], ascending=[True, False]).drop_duplicates(samp_id, keep='first').sort_values('Genome fraction (%)', ascending=False)
     .to_excel(writer, sheet_name=datasheet[1], index=False, na_rep='NA', float_format='%.1f'))
symptoms.to_excel(writer, sheet_name='symptomsDictionary', index=True)
writer.save()

### **County feedback data**

df_counties = df_seq_meta[['CASE_ID', 'sample_name', 'COUNT_RES']]

# unique county identifiers
reports = [('Homabay', 'HCRH'), ('Migori', 'MCRH'), ('Kisii', 'KCRH'), 
           ('Nyamira', 'NCRH'), ('Siaya', 'SCRH'), ('KCSS'), ('Bukavu', 'DRC03'),
           ('Siaya', 'SIAYA'), ('Bungoma', 'BCRH'), ('Kapenguria', 'WPKRH'),
           ('Kitale', 'TNZ'), ('Busia', 'BSCRH'), ('Turkana', 'TCRH'), ('Ethiopia', 'ETH'),
           ('Sudan', 'SDN'), ('Eswatini', 'SWZ'), ('Mauritius', 'MUS'), ('Uganda', 'UGA'),
          ('Angola', 'AGO'), ('UasinGishu', 'UGC'), ('Nyandarua', 'NYA'), ('Bomet', 'BMT'),
          ('Narok', 'NRK'), ('Tharaka-Nithi', 'THN'), ('Nyeri', 'NYR')]
codes = set(df_counties.CASE_ID.map(lambda x: str(x).split('/')[0]))

for report in reports:
    mask1 = df_counties['COUNT_RES'] == report[0]
    mask2 = df_counties['CASE_ID'].str.contains(report[1]) == True
    mask3 = df_counties['CASE_ID'].str.contains(report[0]) == True
#     print(report)
    if len(report) != 2 and report[0] in codes:
        df_report = df_counties[mask3 == True]['sample_name']
        writer = pd.ExcelWriter(f"{glob.glob(f'{parent_dir}/**/CountyFeedbacks', recursive=True)[0]}/{run_dir}_{report}.Analysis.QCstats.xlsx")
    elif len(report) == 2 and report[1] not in codes: 
        pass
    
    elif len(report) == 2 and report[1] in codes:
        df_report = df_counties[mask2 == True]['sample_name']
        writer = pd.ExcelWriter(f"{glob.glob(f'{parent_dir}/**/CountyFeedbacks', recursive=True)[0]}/{run_dir}_{report[0]}.Analysis.QCstats.xlsx")
        for datasheet in datasheets:
            samp_id = 'sample_name' if datasheet[1] != 'summaryReport' else 'Unique lab id'
            (datasheet[0][datasheet[0][samp_id].isin(df_report)].sort_values([samp_id, 'Genome fraction (%)'], ascending=[True, False]).drop_duplicates(samp_id, keep='first')
             .sort_values('Genome fraction (%)', ascending=False)
             .to_excel(writer, sheet_name=datasheet[1], index=False, na_rep='NA', float_format='%.1f'))
        symptoms.to_excel(writer, sheet_name='symptomsDictionary', index=True)
        writer.save()
