==> 05.gisaid_submissions.ipynb <==
## **GISAID Submission Metadata Preparation**

import pandas as pd
import numpy as np
import glob, os, re
from datetime import  datetime

### **Preliminary variables**

#uniqueness in directory and file names is assumed for all analyses
sars_dir = "bioinformatics/github/covid/data/SarsGenomics/SARS-CoV-2"
home_dir = os.getenv('HOME')
parent_dir = glob.glob(f'{home_dir}/{sars_dir}', recursive=True)[0]

### **Load submission form**

df_gisaid = pd.read_excel(glob.glob(f'{parent_dir}/**/20210222_EpiCoV.xls', recursive=True)[0], 'Submissions')

df_gisaid.head()

gisaid_header = ['submitter', 'fn', 'covv_virus_name', 'covv_type', 'covv_passage', 'DT_SAM_COLL', 
 'DT_SAM_RECEP', 'CASE_ID', 'covv_location', 'covv_add_location','covv_host', 'covv_add_host_info', 'covv_sampling_strategy',
 'GEND', 'AGE_YRS', 'covv_patient_status', 'covv_specimen', 'covv_outbreak', 
 'covv_last_vaccinated', 'covv_treatment', 'covv_seq_technology', 'covv_assembly_method', 'covv_coverage', 
 'covv_orig_lab', 'covv_orig_lab_addr', 'S_NUM', 'covv_subm_lab', 'covv_subm_lab_addr', 
 'covv_subm_sample_id', 'covv_authors', 'covv_comment', 'comment_type']

### **Pull-in the metadata**

# get the necessary columns 
df_metadata = pd.read_excel(glob.glob(f'{parent_dir}/**/COVID19-resultsCts-merged-cln.xlsx', recursive=True)[0])[['CASE_ID', 'S_NUM', 'AGE_YRS', 'GEND', 'NAT', 
           'COUNT_RES', 'SAMP_TYPE', 'DT_SAM_COLL', 'DT_SAM_RECEP']]

dt = datetime.today().strftime(format='%d-%m-%Y')
ref = 'seq52'
ref_fasta = 'gisaid_all_12.fasta'

### **Be sure the next input is update**

# import sequence summary data
df_seq_summ = pd.read_excel(glob.glob(f'{parent_dir}/**/seq-summary-metadata_seq52_31-08-2022.xlsx')[0])
df_seq_summ.shape

# mark duplicate
df_seq_summ[df_seq_summ.sequence_name.duplicated(keep='first') == True]

# filter duplicate
df_seq_dedup = df_seq_summ.sort_values(['sequence_name', 'genome_coverage']).drop_duplicates('sequence_name', keep='last')
df_seq_dedup.shape#.head()

# filter sample metadata to sequence metadata
metadata = df_metadata[df_metadata.S_NUM.isin(df_seq_dedup.sequence_name)].rename(columns={'S_NUM': 'sequence_name'})

# mark duplicate
metadata[metadata.sequence_name.duplicated(keep=False) == True]

### **Merge metadata with seq summary data**

# Merge metadata with seq summary data
df_seq_meta = metadata.merge(df_seq_dedup, how='right', on='sequence_name')
df_seq_meta.shape

# mark duplicate
df_seq_meta[df_seq_meta.sequence_name.duplicated(keep=False) == True]

### **Filter non-GISAID samples**

# Filter service samples; remain with internals only
df_seq_filtered1 = df_seq_meta[df_seq_meta.sequence_name.str.contains('COVS') == False]#.head()

# df_seq_filtered2 = df_seq_filtered1[df_seq_filtered1.CASE_ID.str.contains('DRC02') == False]
df_seq_filtered3 = df_seq_filtered1[df_seq_filtered1.sequence_name.str.contains('SSEQ') == False]
df_seq_filtered = df_seq_filtered3  # [df_seq_filtered1.lineage.str.contains('B.1.1.529') == False]
df_seq_filtered.shape[0]

# df_seq_filtered.head()

# update columns
df_gisaid1 = df_seq_filtered.assign(covv_location = 'Africa / ' + df_seq_filtered['NAT'] + ' / ' + df_seq_filtered['COUNT_RES'])
df_gisaid2 = df_gisaid1.assign(GEND = df_gisaid1.GEND.replace(['M', 'F'], ['Male', 'Female']))
df_gisaid3 = df_gisaid2.assign(covv_virus_name = df_gisaid2.sequence_name.map(lambda x: f'hCoV-19/Kenya/ILRI_{x}/'))
df_gisaid4 = df_gisaid3.assign(covv_subm_sample_id = df_gisaid3['sequence_name'])
df_gisaid5 = df_gisaid4.rename(columns={'AGE_YRS': 'covv_patient_age', 'tech': 'covv_seq_technology', 
                                        'sequence_name': 'covv_provider_sample_id', 'GEND': 'covv_gender', 
                                        'SAMP_TYPE': 'covv_specimen', 'DT_SAM_COLL': 'covv_collection_date'})


df_gisaid5.columns

df_subf = df_gisaid5

# assign metada variables
submitter = 'soyola'
fn = ref_fasta
# cvn = 'hCoV-19/Kenya/ILRI_'
vt = 'betacoronavirus'
cp = 'Original'
cal = 'unknown'
ch = 'Human'
cahi = 'unknown'
css = 'Surveillance'
cps = 'unknown'
cps = 'unknown'
# cs = 'NP Swab'
co = 'unknown'
clv = 'unknown'
ct = 'unknown'
# cst = 'Illumina'
cam = 'Consensus'
cc = 'unknown'
col = 'International Livestock Research Institute'
cola = 'Uthiru, Naivasha road, Nairobi-Kenya'
# cpsi = 'ILRI'
csl = 'International Livestock Research Institute'
csla = 'Uthiru, Naivasha road, Nairobi-Kenya'
# cssi = df_subf.index.tolist()
ca = 'Samuel O. Oyola, Gilbert Kibet, Shebbar Osiany, Edward Kiritu, Paul Dobi, Collins Muli, Patrick Amoth, Vishvanath Nene, Benard Mware, Edward O. Abworo'
ccomm = ''
ct = ''

# update gisaid columns
df_subf['submitter'] = submitter
df_subf['fn'] = fn
# df_subf['covv_virus_name'] = cvn
df_subf['covv_type'] = vt
df_subf['covv_passage'] = cp
df_subf['covv_add_location'] = cal
df_subf['covv_host'] = ch
df_subf['covv_add_host_info'] = cahi
df_subf['covv_sampling_strategy'] = css
df_subf['covv_patient_status'] = cps
# df_subf['covv_specimen'] = cs
df_subf['covv_outbreak'] = co
df_subf['covv_last_vaccinated'] = clv
df_subf['covv_treatment'] = ct
# df_subf['covv_seq_technology'] = cst
df_subf['covv_assembly_method'] = cam
df_subf['covv_coverage'] = cc
df_subf['covv_orig_lab'] = col
df_subf['covv_orig_lab_addr'] = cola
# df_subf['covv_provider_sample_id'] = cpsi
df_subf['covv_subm_lab'] = csl
df_subf['covv_subm_lab_addr'] = csla
# df_subf['covv_subm_sample_id'] = cssi
df_subf['covv_authors'] = ca
df_subf['covv_comment'] = ccomm
df_subf['comment_type'] = ct
df_subf.shape

# drop supplimentary columns
df_sub_draft1 = df_subf.drop(['NAT', 'COUNT_RES', 'clade', 'lineage', 'run_num', 'seq_dt', 'lib_prep', 'primer_set',
       'analysis_pipeline', 'sno.'], axis=1)
df_sub_draft1.shape

# df_sub_draft1.head()

# reorder columns according to gisaid template
gisaid_cols = ['submitter', 'fn', 'covv_virus_name', 'covv_type', 'covv_passage',
       'covv_collection_date', 'DT_SAM_RECEP', 'CASE_ID', 'genome_coverage', 'covv_location', 'covv_add_location',
       'covv_host', 'covv_add_host_info', 'covv_sampling_strategy',
       'covv_gender', 'covv_patient_age', 'covv_patient_status',
       'covv_specimen', 'covv_outbreak', 'covv_last_vaccinated',
       'covv_treatment', 'covv_seq_technology', 'covv_assembly_method',
       'covv_coverage', 'covv_orig_lab', 'covv_orig_lab_addr',
       'covv_provider_sample_id', 'covv_subm_lab', 'covv_subm_lab_addr',
       'covv_subm_sample_id', 'covv_authors', 'covv_comment', 'comment_type']


df_sub_draft1['covv_collection_date'].apply(lambda x: pd.Timestamp(x))

df_sub_draft1['covv_collection_date'].astype('datetime64[ns]')

### **Format Dates**

# partition df to with(out) sample collection dates and fortmat dates to strings
df_dt = df_sub_draft1[df_sub_draft1['covv_collection_date'].isna() == False]
df_dt_fmt = df_dt.assign(covv_collection_date=df_dt['covv_collection_date'].astype('datetime64[ns]').dt.strftime('%Y-%m-%d'))
df_missing_dt = df_sub_draft1[df_sub_draft1['covv_collection_date'].isna() == True]

df_missing_dt.head()

# add short/approx. dates for where sample collection dates are missing
df_missing_dt_corr = df_missing_dt.assign(covv_collection_date=df_missing_dt['DT_SAM_RECEP'].apply(lambda x: pd.Timestamp(x).strftime('%Y-%m') if (isinstance(x, pd.Timestamp)) else x.replace(x, pd.NaT)))

# merge approx. dates df to the string-formated dates df
df_sub_draft2 = df_dt_fmt.append(df_missing_dt_corr)
df_sub_draft2.shape

# re-assign the date column with the updated dates column
df_sub_draft2['covv_virus_name'] = df_sub_draft2.covv_virus_name.map(str) + df_sub_draft2.covv_collection_date.map(lambda x: str(x).split('-')[0])

### **Previous Submissions**

# import previous submissions to filter submitted samples
df_prev_sub1 = pd.read_excel(glob.glob(f'{parent_dir}/**/Submissions/20210914_ILRI_gisaid_submission_metadata.xls')[0], sheet_name='Submissions', usecols=['covv_provider_sample_id'])
df_prev_sub2 = pd.read_excel(glob.glob(f'{parent_dir}/**/Submissions/20211102_ILRI_gisaid_submission_metadata.xls')[0], sheet_name='Submissions', usecols=['covv_provider_sample_id'])
df_prev_sub3 = pd.read_excel(glob.glob(f'{parent_dir}/**/Submissions/20211214_ILRI_gisaid_submission_metadata.xls')[0], sheet_name='Submissions', usecols=['covv_provider_sample_id'])
df_prev_sub4 = pd.read_excel(glob.glob(f'{parent_dir}/**/Submissions/20220110_ILRI_gisaid_submission_metadata.xls')[0], sheet_name='Submissions', usecols=['covv_provider_sample_id'])
df_prev_sub5 = pd.read_excel(glob.glob(f'{parent_dir}/**/Submissions/20220204_ILRI_gisaid_submission_metadata.xls')[0], sheet_name='Submissions', usecols=['covv_provider_sample_id'])
df_prev_sub6 = pd.read_excel(glob.glob(f'{parent_dir}/**/Submissions/20220318_ILRI_gisaid_submission_metadata.xls')[0], sheet_name='Submissions', usecols=['covv_provider_sample_id'])
df_prev_sub7a = pd.read_excel(glob.glob(f'{parent_dir}/**/Submissions/20220422_ILRI_gisaid_submission_7a_metadata.xls')[0], sheet_name='Submissions', usecols=['covv_provider_sample_id'])
df_prev_sub7b = pd.read_excel(glob.glob(f'{parent_dir}/**/Submissions/20220422_ILRI_gisaid_submission_7b-sdn_metadata.xls')[0], sheet_name='Submissions', usecols=['covv_provider_sample_id'])


# merge all the previous submissions
df_prev_sub = df_prev_sub1.append([df_prev_sub2, df_prev_sub3, df_prev_sub4, df_prev_sub5, df_prev_sub6, df_prev_sub7a,
                                  df_prev_sub7b])

#df_sub_draft.genome_coverage.to_excel('/home/douso/Documents/test.xlsx')

# filter previously submitted sample
df_sub_draft = df_sub_draft2[df_sub_draft2['covv_provider_sample_id'].isin(df_prev_sub.covv_provider_sample_id) == False].sort_values('covv_provider_sample_id')[gisaid_cols] 

# filter genome coverage/fraction
# export finalised submission form
df_sub_draft_cutoff = df_sub_draft[df_sub_draft.genome_coverage.fillna(np.nan) >= 80.0].drop('genome_coverage', axis=1)

df_sub_draft_cutoff.fillna('unknown').to_excel(f"{glob.glob(f'{parent_dir}/Gisaid')[0]}/gisaid_data_{ref}_{dt}.xlsx", index=False)

# export sample IDs for retreieving and renaming fasta files
df_sub_draft[['covv_provider_sample_id', 'covv_virus_name', 'CASE_ID']].to_csv(f"{glob.glob(f'{parent_dir}/SeqReports')[0]}/All-IDs-Names_{ref}_{dt}.csv", index=False, header=False)
df_sub_draft_cutoff[['covv_provider_sample_id', 'covv_virus_name', 'CASE_ID']].to_csv(f"{glob.glob(f'{parent_dir}/Gisaid')[0]}/gisaid_IDs-Names_{ref}_{dt}.csv", index=False, header=False)

df_sub_draft.shape

# df_final_feedB = pd.read_excel(glob.glob(f'{parent_dir}/**/Submissions/20210914_ILRI_gisaid_submission_metadata_curated.xls')[0], sheet_name='Submissions')

# assess seq platforms used
df_sub_draft['covv_seq_technology'].value_counts()

df_sub_draft.info()
