==> 06.rockefella_report_metrics.ipynb <==
## **Rough work and everything**

import pandas as pd
import seaborn as sns
import numpy as np
import glob, os, re
import matplotlib.pyplot as plt
from datetime import  datetime
from ipywidgets import widgets, interactive

dt = datetime.today().strftime(format='%d-%m-%Y')

### **Establish parent directory**

#uniqueness in directory and file names is assumed for all analyses
sars_dir = "bioinformatics/github/covid/data/SarsGenomics/SARS-CoV-2"
home_dir = os.getenv('HOME')
parent_dir = glob.glob(f'{home_dir}/{sars_dir}', recursive=True)[0]

### **Metadata**

df_meta = pd.read_excel(glob.glob(f'{parent_dir}/**/COVID19-resultsCts-merged-cln.xlsx', recursive=True)[0])
# glob(f'{parent_dir}/**/pc-nc_QC-QA_03-11-2021.xlsx', recursive=True)[0])[['CASE_ID', 'S_NUM']]#, 'G_COV', 'LIN', 'DT_SAM_RECEP', 'DT_SAM_COLL']]

## RF Report August

df_seq_summ = pd.read_excel(glob.glob(f'{parent_dir}/Outputs/seq-summary-metadata_seq52_31-08-2022.xlsx')[0])
df_prev = pd.read_excel(glob.glob(f'{parent_dir}/RFellerReports/rf-20220831_ILRI_gisaid_submission_12_metadata.xls')[0], 
                              sheet_name='Submissions', usecols=['covv_provider_sample_id', 'covv_subm_sample_id'])#covv_subm_sample_id

df_seq_summ. head(2)

df_seq_summ.run_num.unique()

mask1 = df_seq_summ.run_num.isin(['Run52'])#, 'Run41'


latest_seq = df_seq_summ[mask1 == True]
print(f'Total sequenced: {latest_seq.shape[0]}')

mask2, msk1, msk2, msk3 = (latest_seq.case_id.str.contains('SWZ/'), latest_seq.case_id.str.contains('MUS/'),
latest_seq.case_id.str.contains('MUS/'), latest_seq.case_id.str.contains('AGO/'))
# latest_eth = latest_seq[mask2 == True]
# latest_eth.shape[0]#latest_seq.shape[0] - 

print(f'Total submitted: {df_prev.shape[0] - 1}')

mask3 = df_meta.S_NUM.isin(df_prev.covv_subm_sample_id)#covv_subm_sample_id
df_prev.head()

sub_meta = df_meta[mask3 == True]
sub_meta.head(2)

sub_meta.info()

sub_meta[sub_meta.DT_SAM_COLL.isna() == True].shape[0]

sub_fnl = sub_meta[sub_meta.DT_SAM_COLL.isna() == False]


sub_fnl.head(2)#.DT_SAM_COLL.astype('datetime64[ns]')

latest_seq = sub_fnl
dt_delta = sub_fnl[~sub_fnl.CASE_ID.str.contains('MUS/')].DT_SAM_COLL.astype('datetime64[ns]') - pd.Timestamp('2022-08-31')
dt_delta1 = sub_fnl[sub_fnl.CASE_ID.str.contains('MUS/')].DT_SAM_COLL.astype('datetime64[ns]') - pd.Timestamp('2022-08-31')
# dt_delta2 = sub_fnl[msk2].DT_SAM_COLL.astype('datetime64[ns]') - pd.Timestamp('2022-06-02')
# dt_delta3 = sub_fnl[msk3].DT_SAM_COLL.astype('datetime64[ns]') - pd.Timestamp('2022-06-02')
print(f'Local: {dt_delta.mean()}')
print(f'Mauritius: {dt_delta1.mean()}')
# print(f'Uganda: {dt_delta2.mean()}')
# print(f'Angola: {dt_delta3.mean()}')
